src/advanced_perception.py
import numpy as np
import tensorflow as tf
import concurrent.futures
from spectral import *
import google.generativeai as genai
from raman_libs import RamanSpectrometer  # Hypothetical Raman spectroscopy library
from gas_sensor_array import GasSensorArray  # Hypothetical gas sensor library

class AdvancedPerceptionSystem:
    def __init__(self, config_manager, event_system):
        self.config = config_manager
        self.events = event_system
        
        # Initialize Gemini AI
        genai.configure(api_key=self.config.get('ai', 'gemini_key'))
        self.model = genai.GenerativeModel('gemini-pro')
        
        # Initialize sensor arrays
        self.gas_sensor = GasSensorArray()
        self.raman_spectrometer = RamanSpectrometer()
        self.hyperspectral_camera = None
        
        # Initialize neural networks
        self.material_classifier = self._load_material_classifier()
        self.density_estimator = self._load_density_estimator()
        self.state_classifier = self._load_state_classifier()
        
        # Initialize thread pool for parallel processing
        self.thread_pool = concurrent.futures.ThreadPoolExecutor(max_workers=8)
        
    def analyze_environment(self, sensor_data):
        """Parallel processing of environmental data"""
        try:
            # Create futures for parallel processing
            futures = {
                'material': self.thread_pool.submit(self._analyze_material, sensor_data),
                'state': self.thread_pool.submit(self._analyze_state, sensor_data),
                'density': self.thread_pool.submit(self._analyze_density, sensor_data),
                'gas': self.thread_pool.submit(self._analyze_gases, sensor_data),
                'hazards': self.thread_pool.submit(self._analyze_hazards, sensor_data)
            }
            
            # Gather results
            results = {
                key: future.result() for key, future in futures.items()
            }
            
            # Use Gemini for complex reasoning about the environment
            ai_analysis = self._ai_enhanced_analysis(results)
            results['ai_insights'] = ai_analysis
            
            return results
            
        except Exception as e:
            print(f"Environment analysis error: {e}")
            return None
            
    def _analyze_material(self, sensor_data):
        """Analyze material composition using multiple techniques"""
        try:
            # Raman spectroscopy analysis
            raman_data = self.raman_spectrometer.analyze(sensor_data['spectral'])
            
            # Hyperspectral imaging analysis
            hyperspectral_data = self._process_hyperspectral(sensor_data['hyperspectral'])
            
            # Neural network classification
            material_properties = self.material_classifier.predict({
                'raman': raman_data,
                'hyperspectral': hyperspectral_data,
                'visual': sensor_data['visual']
            })
            
            return material_properties
            
        except Exception as e:
            print(f"Material analysis error: {e}")
            return None
            
    def _analyze_state(self, sensor_data):
        """Determine physical state (solid, liquid, gas, plasma)"""
        try:
            # Combine multiple sensor inputs for state detection
            state_features = {
                'thermal': sensor_data['thermal'],
                'spectral': sensor_data['spectral'],
                'visual': sensor_data['visual'],
                'em_field': sensor_data['electromagnetic']
            }
            
            # Neural network classification of state
            state_probabilities = self.state_classifier.predict(state_features)
            
            # Additional plasma detection using electromagnetic sensors
            plasma_detection = self._detect_plasma(sensor_data['electromagnetic'])
            
            return {
                'state_probabilities': state_probabilities,
                'plasma_detected': plasma_detection
            }
            
        except Exception as e:
            print(f"State analysis error: {e}")
            return None
            
    def _analyze_density(self, sensor_data):
        """Estimate density of materials"""
        try:
            # Combine acoustic and pressure sensor data
            density_features = {
                'acoustic': sensor_data['acoustic'],
                'pressure': sensor_data['pressure'],
                'visual_depth': sensor_data['depth'],
                'material_interaction': sensor_data['interaction']
            }
            
            # Neural network estimation of density
            density_estimate = self.density_estimator.predict(density_features)
            
            return density_estimate
            
        except Exception as e:
            print(f"Density analysis error: {e}")
            return None
            
    def _analyze_gases(self, sensor_data):
        """Analyze gas composition and properties"""
        try:
            # Get gas sensor array data
            gas_data = self.gas_sensor.get_readings()
            
            # Analyze for specific hazardous gases
            gas_composition = self.gas_sensor.analyze_composition()
            
            # Check for corrosive properties
            corrosion_risk = self._assess_corrosion_risk(gas_composition)
            
            return {
                'composition': gas_composition,
                'corrosion_risk': corrosion_risk,
                'concentration': gas_data['concentration'],
                'flow_rate': gas_data['flow_rate']
            }
            
        except Exception as e:
            print(f"Gas analysis error: {e}")
            return None
            
    def _ai_enhanced_analysis(self, sensor_results):
        """Use Gemini for complex reasoning about the environment"""
        try:
            # Prepare context for AI
            context = self._prepare_ai_context(sensor_results)
            
            # Get AI analysis using Gemini
            response = self.model.generate_content(
                prompt=f"""You are an advanced material analysis AI. 
                Analyze the following sensor data and provide insights about the environment 
                and potential hazards:
                
                {context}"""
            )
            
            return response.text
            
        except Exception as e:
            print(f"AI analysis error: {e}")
            return None

# If you need to export additional classes/functions
__all__ = ['AdvancedPerceptionSystem']